Passive reflex mode with multiple argument support
This is a 2nd pass at adding passive reflex support. I took pains to follow the general structure of what you had, and this functionality is now activated when the register() call is made from the Stimulus controller. Arguments are any data attribute that starts with reflex-arg and they are processed in the order they appear.