update github action triggers
Update the triggers so this only runs on pull request triggers.

Currently, if a user has commit access to a repo, the lint actions will run twice. This is because GitHub triggers to events.

1. There is a push event on the repo. Only users with commit access can trigger these.
2. There is a pull request event on the repo. As it stands, this seems to only trigger on pushes to the repo.

Triggering only on pull requests seems like the best option here. This will allow checks to run on any commits to a PR, and not run checks twice for a user with commit access.