Allow #stimulate to use promises
The basic idea here is that we keep a table of promises indexed by
randomly generated UUIDs. The UUIDs are sent to the server on each
reflex call and are returned with responses, and thus can be used to
look up the promise we kept in our table, which we then resolve or
reject. This allows us to do stuff like

  this.stimulate('MyReflex#action')
    .then(() => this.doSomething())
    .catch(() => this.handleError())