Do not run sanity check on `rails generate stimulus_reflex:config`
If a sanity check fails, the user gets a hint to create the configuration file with
`rails generate stimulus_reflex:config`, however this command will also fail because it
will also run the sanity check. Therefore we have to bypass the sanity check if
the caller is a `generate_command`.

Co-authored-by: Marco Roth <marco.roth@intergga.ch>