more explicit quick_start doc increment example
Hi and thank you for a super nice approach. 

I'm just learning discovering Stimulus Reflex, and while going through the quick start example, I found it slightly complicated to understand. I had a hard time understanding how instance variables set inside Reflexes would become available in the view, and how it interacted with controller set instance variables.

I think part of this blur came from the fact that your example currently relies on the fact that `nil.to_i == 0` which is not super obvious or expected. I'm suggesting a slightly more explicit code that explicitly tests for the presence of the instance variable in the view and then does something. I hope it may help future readers to understand that the variable will only be set when coming from a reflex, and not from a regular page load.  

I had to watch the nice and short screencast to actually understand this. Maybe it'd be nice to actually a small piece of doc about this interaction with instance variable set in controllers, but I'll leave that up to you!