Updated gemspec to specific Rails gems
Most apps don't require the full suite of Rails gems, and can avoid
installing them to save CI and deploy time. This gem, however, had a
dependency on the Rails meta-gem which ended up installing them all
anyways.

Instead, we can limit it to only the Rails gems that we directly depend upon.