Make `ReflexData#version` more error-proof
Sometimes when you are using a locally linked version of the JavaScript
package `data["version"]` will be `nil`. The sanity check relies on this
method to figure out if the version matches. But since the version is
`nil` in this case the `gsub` call will fail, resulting in the whole
sanity check to crash with an `undefined method `gsub' for nil:NilClass`
error.

This commit makes that method more error prove so that the sanity check
can actually you that the version didn't match. Or if you ignored the
mismatch that it can continue with the process.