Don't add `data-controller` to reflex element if parent already holds instance (#636)
There's was an edge case where we would add another `data-controller`
attribute to the reflex element even if a parent element already had a
matching controller instance.

This markup: 
```html
<div data-controller="example">
  <button 
    data-reflex="click->Example#increment"
  >Click</button>
</div>
```
would get transformed to:
```diff
<div data-controller="example">
  <button 
    data-reflex="click->Example#increment" 
+   data-controller="example" 
+   data-action="click->example#__perform"
  >Click</button>
</div>
```
Which ends up with two controller instances of `example` even if we
actually just needed one.

This pull request optimizes this by just conditionally adding the
`example` controller to the `data-controller` attribute on the reflex
element.

Now with this pull request the end-result is just:
```diff
<div data-controller="example">
  <button 
    data-reflex="click->Example#increment" 
+   data-action="click->example#__perform"
  >Click</button>
</div>
```

## Why should this be added

We shouldn't create more controller instances than necessary. In this
case it was apparent because the `connect()` callback was fired twice.
To not fire the method twice we add the condition.