Improve DX around version mismatch warnings (#641)
This pull request enhances the developer experience when the
StimulusReflex client- and server-side versions mismatch. It will show a
toast message in the lower right corner of the screen if there's a
version mismatch. Previously it would just show it in the Rails server
log.

It will only show the toasts if you try to trigger a reflex action, it
won't complain earlier, which is also much friendlier, because in
previous versions of StimulusReflex it wouldn't even let you start the
Rails server.

This is what happens now if you try to trigger a reflex:

* on `:exit` it will show a red error toast message and won't execute
the trigged reflex:
![Screenshot 2023-03-13 at 15 43
43](https://user-images.githubusercontent.com/6411752/224736409-aac0d8bf-9d11-4dff-82ae-b9d7b0926b6b.png)

* on `:warn` it will show a yellow warning toast message but will still
execute the triggered reflex:
![Screenshot 2023-03-13 at 15 43
28](https://user-images.githubusercontent.com/6411752/224735908-4b56f4ad-9537-48cc-8883-959cdafd399a.png)


* on `:ignore` it won't show anything and also will also regularly
execute the triggered reflex.

Note: the `toastify-js` dependency will be bundled up with the built
JavaScript version of the package, so it won't show up as a runtime
dependency in applications running the `stimulus_reflex` npm package.

The toasts are currently not disabled in production environments, but
I'm also not sure if we should. Any opinions on this?

## Why should this be added

Improves the developer experience and makes it easier to noticed/work
with.