Don't choose first controller if no matching StimulusReflex-enabled controller was found (#670)
This pull request solves an edge-case for scenarios when StimulusReflex
tries to detect the right controller to use for performing the
StimulusReflex `__perform` controller action for executing reflexes.

For example, if an element had a `data-controller` and a `data-reflex`
attribute, and the controller in `data-controller` was a
StimulusReflex-enabled controller, but didn't match the Reflex specified
in the `data-reflex` there were some cases where it wouldn't add the
default `stimulus-reflex` controller to the element, but would still
reference the `stimulus-reflex` controller in the `data-action`
attribute. This basically lead to a no-op for the reflex action.

Like:
```html
<a
  data-controller="
    some-sr-controller
  "
  data-reflex="click->Example#doSomething"
></a>
```

was expanded to this in some cases:
```diff
<a
  data-controller="
    some-sr-controller
  "
  data-reflex="click->Example#doSomething"
+ data-action="click->stimulus-reflex#__perform"
></a>
```

Whereas it should have been:
```diff
<a
  data-controller="
    some-sr-controller
+   stimulus-reflex
  "
  data-reflex="click->Example#doSomething"
+ data-action="click->stimulus-reflex#__perform"
></a>
```
With this, the `data-action` is not referencing an undefined Stimulus
controller anymore and the reflex should execute again.

## Why should this be added

For consitancy reasons, so that the declarative reflex syntax still
continues to make sense and work as expected.

Thanks to @foliosus for helping to debug this!