Fix exception in ActionCable channel due to reflexes overwriting each other's data (#663)
Think of a page with multiple StimulusReflex controllers hooked up to
the `stimulus-reflex:connected` event and the markup (Haml in this
example) looking something like this:

```haml
%div{ data: { controller: "sandwich", action: "stimulus-reflex:connected@document->sandwich#eat" } }

%div{ data: { controller: "tea", action: "stimulus-reflex:connected@document->tea#drink" } }
```

The actions are backed by simple JavaScript controllers and reflex
actions looking as follows. The reflex actions have different arity: one
take no arguments, the other takes one.

```js
// sandwich.js:

eat() {
  this.stimulate("Sandwich#eat");
}

// tea.js:

drink() {
  this.stimulate("Tea#drink", "green");
}
```

```ruby
# sandwich_reflex.rb:

class SandwichReflex < ApplicationReflex
  def eat
    # ...
  end
end

# tea_reflex.rb:

class TeaReflex < ApplicationReflex
  def drink(flavour)
    # ...
  end
end
```

This kind of setup creates a situation where
`StimulusReflex::Channel#receive` is invoked twice in a very short
period of time. And since both messages are processed by the **same
instance** of the ActionCable channel (as can be verified by printing
out `self.object_id` at the beginning of the method), the use of an
instance variable (`@reflex_data`) in the current implementation leads
to a situation where it gets overrwritten by the second message before
the first one has been processed. As a result, `reflex_data` inside
`#delegate_call_to_reflex` ends up containing the same set of data
which, after being checked against the method signature of both of the
reflex actions, raises an `ArgumentError` due to arity mismatch.

This pull request fixes the issue by eliminating the use of instance
variables inside `StimulusReflex::Channel#receive` in favour of bundling
the message payload with the respective instance of the reflex. It also
simplifies the reflex factory and constructor by delegating most of the
calls to the `StimulusReflex::ReflexData` instance.

## Why should this be added

In modern web applications, it is common to load parts of the page
asynchronously for a more responsive user experience. With
StimulusReflex, it is most easily achieved by hooking up to the
`stimulus-reflex:connected` event. This pull request fixes an exception
that prevented the use of the specified event with multiple reflexes.