Explicitly require `ostruct` for `StimulusReflex::Dataset` and `Element` (#697)
Rubocop v1.63.0 updates json v2.7.2. StimulusReflex relies on OpenStruct in both the
Dataset and the Element models. Attempting to run rails, ie a rails console, results in the error. This pull requets fixes this by explicitly requiring `ostruct`.

We should think about migrating off of `OpenStruct`, but this PR should make it work for now as a work-around.

Fixes #695