Isolate and namespace `StimulusReflex::Installer` (#699)
This pull request isolates the methods defined
`lib/stimulus_reflex/installer.rb` into a `StimulusReflex::Installer`
constant so they don't pollute the global namespace if the file is
required.

CableReady got a similar PR:
https://github.com/stimulusreflex/cable_ready/pull/295

## Why should this be added

Exposing all the methods from the installer in the global namespace
could lead to issues down the road. Methods like `options` are likely to
be overridden.