Require Thor Dependency
Applications using StimulusReflex were failing
during deploys because of the Thor dependency
being referenced but not listed as a dependency.
This commit adds the require statement to the area
of the code that uses it but does not add it as a
dependency due to Thor being a dependency of
railties already.