Add page which describes the "redisless" setup (#703)
This PR adds a new page which describes how to setup StimulusReflex
without Redis. I extracted it into a separate page because I felt it was
too much to rewrite all the existing documentation, which relies on
Redis. I assume most people will use Redis, and the "redisless"
configuration likely is a niche thing.

Fixes #684