Declare dependency on ostruct in gemspec (#735)
The absence of this declaration in `stimulus_reflex.gemspec` was
emitting this warning in projects using this library:

> warning: .rubies/ruby-3.4.2/lib/ruby/3.4.0/ostruct.rb was loaded from
the standard library, but will no longer be part of the default gems
starting from Ruby 3.5.0. You can add ostruct to your Gemfile or gemspec
to silence this warning. Also please contact the author of
stimulus_reflex-3.5.3 to request adding ostruct into its gemspec.


Related to #695, #698 

## Why should this be added
I know that the long-term goal is to move away from using ostruct
entirely, but since it's still used in the meantime, this feels like a
reasonable thing to do.